Added Hot Reloading functionality for the application.
If a change is made while the development server is running, the app will try to hot update without reloading and if the update was unaccepted or failed, the app will reload the page(this is configured using a 'reload' query param in the webpack module 'config.entry' value).

Also updated the 'webpack-cli' package as an older version of it was being used to avoid erros when running the 'webpack' command earlier.