Enhanced the Redux Mithril binding 'connect' function through Route Resolvers.
In the previous, the store was subscribed inside the 'connect' function. It was not ideal to implement other possible Route Resolver functionalities (e.g. an application layout component) as the 'connect' function was returning a Route Resolver to deliver updated vnode attrs to the component on store update. So a new function 'RouteResolver' was created to return a Route Resolver object. This function can handle connected & non connected components as params and returns the components along with its vnode attrs through a Route Resolver object.

The connect function now solely handles getting latest props from the store & is not concerned with store subscriptions. The store is subscribed in the Redux config file (src/reduxConfig/config.js) and m.redraw() is called in the subscription callback. A more efficient way of redrawing rather than calling it on every dispatch call may be implemented later.